This is a seriously large checkin to fix a pretty serious bug.
Bing Maps has an unfortunate bug which causes crashes while navigating between pages. The code is outside of our control and there is nothing we can do to stop it.  The bug occurs when you navigate quickly between pages. Eventually, a NavigationFailed event is sent to the UI thread using Dispatcher.Invoke.  The exception always re-enters into native code that is outside of our control and it throws an E_UNEXPECTED COM error which crashes the application.

The only solution to this issue is to change navigation so that instead of moving between pages, we simply change the control on the left and alter the databinding. This way we only ever have one map control in the app and there are no actual page navigations, just the appearance of it.

This also has a number of nice side effects:

1) The map is much more fluid because it does not refresh from scratch when you go to a new page.
2) We use less data because we are not reloading the same map over and over again.
3) Page navigation between favorites and real time is cleaner.
4) The app appears to run much faster now.

There will be bugs as a result of this change (for instance, the help page doesn't work right now - we will need to discuss that), but the original bug is fixed and will not repro with this change.

git-tfs-id: [https://mbraude.visualstudio.com:443/DefaultCollection]$/Windows8/OneBusAway;C148