Adding snap view support!
Since WinRT doesn't include triggers in their API, I invented my own :). The TriggerManager class dfines an attached property that returns a list of Triggers. When the XAML parser runs, it adds triggers to this list. When an item is added to the list, we have a control and a Trigger object that the control is bound to.

So all we do is wire up the trigger's Value property to the binding. This way when the bound value changes, it triggers the changed handler on the Value property, which gives us the opportunity to change states!

The rest of this changeset is just adding the right xaml tags to support the triggers and the animations between Visual States. I am also adding a ScalableTextBlock control that will automatically change the font size depending on the size of the screen.

git-tfs-id: [https://mbraude.visualstudio.com:443/DefaultCollection]$/Windows8/OneBusAway;C156